US airport network, 2010 December

The network of passanger flights between airports in the United States. The data set was compiled based on flights in 2010 December. This network is directed and edge directions correspond to flight directions. Each edge is specific to a single carrier aircraft type. Multiple carriers between the same two airports are denoted by multiple edges.

See information about the included meta-data below.

USairports

Format

A directed igraph graph object, with multiple edges. It has a ‘name’ graph attribute, and several vertex and edge attributes. The vertex attributes:

name

Symbolic vertex name, this is the three letter IATA airport code.

City

City and state, where the airport is located.

Position

Position of the airport, in WGS coordinates.

Edge attributes:

Carrier

Name of the airline. The network includes both domestic and international carriers that performed at least one flight in December of 2010.

Departures

The number of departures (for a given airline and aircraft type.

Seats

The total number of seats available on the flights carried out by a given airline, using a given aircraft type.

Passengers

The total number of passangers on the flights carried out by a given airline, using a given aircraft type.

Aircraft

Type of the aircraft.

Distance

The distance between the two airports, in miles.

Source

Most of this information was downloaded from The Research and Innovative Technology Administration (RITA). See http://www.rita.dot.gov/about_rita/ for details. The airport position information was collected from Wikipedia and other public online sources.