Zachary's karate club network

Social network between members of a university karate club, led by president John A. and karate instructor Mr. Hi (pseudonyms).

The edge weights are the number of common activities the club members took part of. These activities were:

Association in and between academic classes at the university.

Membership in Mr. Hi's private karate studio on the east side of the city where Mr. Hi taught nights as a part-time instructor.

Membership in Mr. Hi's private karate studio on the east side of the city, where many of his supporters worked out on weekends.

Student teaching at the east-side karate studio referred to in (2). This is different from (2) in that student teachers interacted with each other, but were prohibited from interacting with their students.

Interaction at the university rathskeller, located in the same basement as the karate club's workout area.

Interaction at a student-oriented bar located across the street from the university campus.

Attendance at open karate tournaments held through the area at private karate studios.

Attendance at intercollegiate karate tournaments held at local universities. Since both open and intercollegiate tournaments were held on Saturdays, attendance at both was impossible.

Zachary studied conflict and fission in this network, as the karate club was split into two separate clubs, after long disputes between two factions of the club, one led by John A., the other by Mr. Hi.

The ‘Faction’ vertex attribute gives the faction memberships of the actors. After the split of the club, club members chose their new clubs based on their factions, except actor no. 9, who was in John A.'s faction but chose Mr. Hi's club.

karate

Format

An undirected igraph graph object. Vertex no. 1 is Mr. Hi, vertex no. 34 corresponds to John A.

Graph attributes: ‘name’, ‘Citation’, ‘Author’.

Vertex attributes: ‘name’, ‘Faction’, ‘color’ is the same as ‘Faction’, ‘label’ are short labels for plotting.

Edge attribute: ‘weight’.

Source

See reference below.

References

Wayne W. Zachary. An Information Flow Model for Conflict and Fission in Small Groups. Journal of Anthropological Research Vol. 33, No. 4 452-473