Visuotactile brain areas and connections

Graph model of the visuotactile brain areas and connections of the macaque monkey. The model consists of 45 areas and 463 directed connections.

macaque

Format

A directed igraph graph object with vertex attributes ‘name’ and ‘shape’.

This dataset is licensed under a Creative Commons Attribution-Share Alike 2.0 UK: England & Wales License, see http://creativecommons.org/licenses/by-sa/2.0/uk/ for details. Please cite the reference below if you use this dataset.

Source

See reference below.

References

Negyessy L., Nepusz T., Kocsis L., Bazso F.: Prediction of the main cortical areas and connections involved in the tactile function of the visual cortex by network analysis. European Journal of Neuroscience, 23(7): 1919-1930, 2006.