Yeast protein interaction network

Comprehensive protein-protein interaction maps promise to reveal many aspects of the complex regulatory network underlying cellular function.

This data set was compiled by von Mering et al. (see reference below), combining various sources. Only the interactions that have ‘high’ and ‘medium’ confidence are included here.

yeast

Format

An undirected igraph graph object. Its graph attributes: ‘name’, ‘Citation’, ‘Author’, ‘URL’. ‘Classes’. The ‘Classes’ attribute contain the key for the classification labels of the proteins, in a data frame, the original MIPS categories are given after the semicolon:

E

energy production; energy

G

aminoacid metabolism; aminoacid metabolism

M

other metabolism; all remaining metabolism categories

P

translation; protein synthesis

T

transcription; transcription, but without subcategory ‘transcriptional control’

B

transcriptional control; subcategory ‘transcriptional control’

F

protein fate; protein fate (folding, modification, destination)

O

cellular organization; cellular transport and transport mechanisms

A

transport and sensing; categories ‘transport facilitation’ and ‘regulation of / interaction with cellular environment’

R

stress and defense; cell rescue, defense and virulence

D

genome maintenance; DNA processing and cell cycle

C

cellular fate / organization; categories ‘cell fate’ and ‘cellular communication / signal transduction’ and ‘control of cellular organization’

U

uncharacterized; categories ‘not yet clear-cut’ and ‘uncharacterized’

Vertex attributes: ‘name’, ‘Description’, ‘Class’, the last one contains the class of the protein, accoring to the classification above.

Note that some proteins in the network did not appear in the annotation files, the ‘Class’ and ‘Description’ attributes are NA for these.

Source

The data was downloaded from http://www.nature.com/nature/journal/v417/n6887/suppinfo/nature750.html.

References

Comparative assessment of large-scale data sets of protein-protein interactions. Christian von Mering, Roland Krause, Berend Snel, Michael Cornell, Stephen G. Oliver, Stanley Fields and Peer Bork. Nature 417, 399-403 (2002)